Reorganized header/footer elements, created Footer
