+++ src/components/header-footer/Footer.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+export default function Footer() {
+  return (
+    <div className="footer-wrapper">
+      <div className="footer">
+        <p>The Backfield v1.025.1026.1</p>
+      </div>
+    </div>
+  );
+}
